refactor: Position BigDecimal 교체
- Position의 double 속성을 BigDecimal로 교체
- 관련 테스트 및 코드 수정

+++ src/main/java/com/atwoz/member/application/info/dto/profile/PositionWriteRequest.java
@@ -1,12 +1,13 @@
 package com.atwoz.member.application.info.dto.profile;
 
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record PositionWriteRequest(
         @NotNull(message = "위도가 있어야 합니다.")
-        Double latitude,
+        BigDecimal latitude,
 
         @NotNull(message = "경도가 있어야 합니다.")
-        Double longitude
+        BigDecimal longitude
 ){
 }
